backend/pdf-service-4/requirements.txt
Flask
WeasyPrint
Flask-CORS
gunicorn
google-cloud-storage